Use try_into instead of shifting per byte
I haven't actually looked at the generated code, but this feels
like it should be significantly more optimizable than the iterative
version. I haven't found a way to extend this to the integer variant
yet due to their flexibility.

This raises the minimum rust version to 1.40.0 (in case it wasn't above
that already).